solved task 2 check out
