+++ js-modules/js-structure_shapes/components/Circle/Circle.js
@@ -1,0 +1,9 @@
+import { getRandomColor } from "../../utils/randomColor.js";
+
+const circle = document.createElement("div");
+circle.classList.add("circle");
+circle.addEventListener("click", () => {
+  const randomColor = getRandomColor();
+  circle.style.backgroundColor = randomColor;
+});
+export default circle;
