+++ js-modules/js-structure_shapes/components/Pentagon/Pentagon.js
@@ -1,0 +1,9 @@
+import { getRandomColor } from "../../utils/randomColor.js";
+
+const pentagon = document.createElement("div");
+pentagon.classList.add("pentagon");
+pentagon.addEventListener("click", () => {
+  const randomColor = getRandomColor();
+  pentagon.style.backgroundColor = randomColor;
+});
+export default pentagon;
