+++ js-modules/js-structure_shapes/components/Square/Square.js
@@ -1,0 +1,9 @@
+import { getRandomColor } from "../../utils/randomColor.js";
+
+const square = document.createElement("div");
+square.classList.add("square");
+square.addEventListener("click", () => {
+  const randomColor = getRandomColor();
+  square.style.backgroundColor = randomColor;
+});
+export default square;
